Move prefect tasks for osm_features into osm_features_tasks.
Move functions to access spatialite and run prefect flows into separate utils file
